backend/pyproject.toml
[project]
name = "watercrawl"
version = "0.12.1"
description = "WaterCrawl is a smart and extensible web crawler that extracts structured knowledge from websites, generates LLM-ready Markdown files, and visualizes crawl strategies. Built with Django, Scrapy, and Playwright."
authors = [
    {name = "AmirMohsen Asaran",email = "amir.asaran92@gmail.com"}
]
readme = "../README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "gunicorn (==23.0.0)",
    "django (==5.2.9)",
    "django-environ (==0.11.2)",
    "scrapy (==2.13.4)",
    "djangorestframework (==3.15.2)",
    "celery (==5.4.0)",
    "redis (==5.2.1)",
    "django-celery-beat (==2.8.1)",
    "django-celery-results (==2.5.1)",
    "drf-spectacular (==0.28.0)",
    "drf-spectacular-sidecar (==2024.12.1)",
    "psycopg2-binary (==2.9.10)",
    "html2text (==2024.2.26)",
    "scrapy-splash (==0.9.0)",
    "django-minio-backend (==3.8.0)",
    "djangorestframework-simplejwt (==5.5.1)",
    "django-cors-headers (==4.6.0)",
    "watercrawl-plugin (==0.1.0)",
    "watercrawl-openai (==0.1.1)",
    "httpx (==0.28.1)",
    "stripe (==11.5.0)",
    "django-filter (==25.1)",
    "django-redis (==5.4.0)",
    "sentry-sdk[aiohttp,celery,django] (>=2.43.0,<3.0.0)",
    "urllib3 (>=2.6.0)",
    "django-redis (==5.4.0)",
    "pypandoc-binary (>=1.15,<2.0)",
    "opensearch-py (>=3.0.0,<4.0.0)",
    "weaviate-client (>=4.9.0,<5.0.0)",
    "watercrawl-openai (>=0.1.1,<0.2.0)",
    "adrf (>=0.1.11,<0.2.0)",
    "langchain (>=1.0.8,<2.0.0)",
    "langchain-openai (>=1.0.3,<2.0.0)",
    "langchain-google-genai (>=4.0.0,<5.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "langchain-text-splitters (>=1.0.0,<2.0.0)",
    "mcp (>=1.22.0,<2.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "pgvector (>=0.4.2,<0.5.0)",
    "cohere (>=5.20.0,<6.0.0)",
    "langchain-cohere (>=0.5.0,<0.6.0)",
    "langchain-anthropic (>=0.3.0,<0.4.0)",
    "boto3 (>=1.35.0,<2.0.0)",
    "drf-nested-routers (>=0.95.0,<0.96.0)",
    "langchain-ollama (>=1.0.1,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"
ruff = "^0.11.2"


[tool.poetry.dependencies]
[tool.ruff]
exclude = ["migrations", "__pycache__", "venv"]
